// src/notifications/enums/notification.enum.ts
export enum ENotificationType {
  USER_VOTED_POST = 'USER_VOTED_POST',
  USER_COMMENTED_POST = 'USER_COMMENTED_POST',
  USER_REACTED_POST = 'USER_REACTED_POST',
  USER_SHARED_POST = 'USER_SHARED_POST',

  USER_LIKED_COMMENT = 'USER_LIKED_COMMENT',
  USER_REPLIED_COMMENT = 'USER_REPLIED_COMMENT',

  USER_FOLLOWED_USER = 'USER_FOLLOWED_USER',

  CUSTOM = 'CUSTOM',
}